factor::numeric: Move gcd (and its test) to a submodule
